% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_maker.R
\name{map_maker_random}
\alias{map_maker_random}
\title{Make a random map}
\usage{
map_maker_random(
  n_antigens,
  n_sera,
  range,
  dimensions = 2,
  antigen_density = n_antigens/range^dimensions,
  rdistribution = stats::runif,
  coincident = T,
  seed
)
}
\arguments{
\item{n_antigens}{The number of antigen points}

\item{n_sera}{The number of serum points}

\item{range}{The maximum range of the map (in all dimensions)}

\item{dimensions}{The number of dimensions the map}

\item{antigen_density}{The density of antigen points, which can be used to indirectly specify the range. Only used when range is not specified}

\item{rdistribution}{A function that generates random numbers for the spatial distribution of points. The default is runif, but can be another distribution or a user generated function.}

\item{coincident}{TRUE if the serum points are in the same position as the antigen points, FALSE otherwise}

\item{seed}{Random seed}
}
\value{
list
}
\description{
Create random map positions and return the positions and distance matrix.
}
\examples{

map_maker_random(5, 5, 10)
}
