% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_to_titre.R
\name{noise_dist_to_hi_titre}
\alias{noise_dist_to_hi_titre}
\title{Convert both true and noisy antigenic distances to HI titres}
\usage{
noise_dist_to_hi_titre(
  noise_dists,
  base = 2,
  divisor = 10,
  max_log_titre = 9,
  min_log_titre = 0
)
}
\arguments{
\item{noise_dists}{The output from add_noise}

\item{base}{The base that should be used for converting distances to titres. The default is base 2}

\item{divisor}{The numerator for the conversion. This is the number you divide the HI titre by when converting to an antigenic distance. The default is 10.}

\item{max_log_titre}{The highest logged titre possible. This can be the same for all sera, or a vector that is recycled.}

\item{min_log_titre}{The smallest logged titre possible. As above.}
}
\value{

}
\description{
A wrapper for dist_to_hi_titre that converts the true antigenic distances and the noisy antigenic distances, as output from add_noise
}
\examples{
m <- map_maker_random(5, 5, 10)
n <- add_noise(m$dist)
noise_dist_to_hi_titre(n)
}
