% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_to_titre.R
\name{dist_to_hi_titre}
\alias{dist_to_hi_titre}
\title{Converting antigenic distances to titres}
\usage{
dist_to_hi_titre(map, dists, base = 2, divisor = 10)
}
\arguments{
\item{map}{an antigenic map generated by a map_maker function}

\item{dists}{a matrix of antigenic distances}

\item{base}{the base that should be used for converting distances to titres. The default is base 2}

\item{divisor}{the numerator for the conversion. This is the number you divide the HI titre by when converting to an antigenic distance.}
}
\value{

}
\description{
Takes an antigenic distance matrix and converts into values like HI titres. Antigenic distances are calculated by taking the logarithm of the HI titres, so the reverse is done here (exponentiated). So that the antigenic distances are generally whole numbers when calculated from standard HI tables, certain defaults are used.
}
\examples{
m <- map_maker_random(5, 5, 10)
dist_to_hi_titre(m)
}
