% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_maker.R
\name{map_maker_coord}
\alias{map_maker_coord}
\title{Make a map with set coordinates for each antigen & serum}
\usage{
map_maker_coord(
  n_antigens,
  n_sera,
  true_ag_coord,
  true_sr_coord = true_ag_coord,
  range,
  dimensions = 2,
  antigen_density = n_antigens/range^dimensions,
  rdistribution = stats::runif,
  seed
)
}
\arguments{
\item{n_antigens}{The number of antigen points for each cluster}

\item{n_sera}{The number of serum points for each cluster}

\item{true_ag_coord}{The coordinates of antigen points for each cluster. True coordinates need to be specified for the maximum number of dimensions.}

\item{true_sr_coord}{The coordinates of serum points for each cluster. True coordinates need to be specified for the maximum number of dimensions.}

\item{range}{The maximum range of each cluster (in all dimensions)}

\item{dimensions}{The number of dimensions the map}

\item{antigen_density}{The density of antigen points, which can be used to indirectly specify the range for a cluster. Only used when range is not specified}

\item{rdistribution}{A function that generates random numbers for the spatial distribution of points. The default is runif, but can be another distribition or a user generated function.}

\item{seed}{Random seed}
}
\value{
list
}
\description{
Create random map positions for specified cluster positions and numbers
}
\examples{

map_maker_coord(5, 5, matrix(c(1,0,0,0,1,1,1,0,0,0), ncol=2, byrow=TRUE), range=10)
}
