% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_maker.R
\name{map_maker_random}
\alias{map_maker_random}
\title{Make a random map}
\usage{
map_maker_random(
  n_antigens = 10,
  n_sera = 10,
  range = c(0, 10),
  dimensions = 2
)
}
\arguments{
\item{n_antigens}{The number of antigen points}

\item{n_sera}{The number of serum points}

\item{range}{The range of the map (in all dimensions)}

\item{dimensions}{The number of dimensions the map}
}
\value{

}
\description{
Create random map positions and return the positions and distance matrix. Currently produces a map as a line, square, cube or hypercube
}
\examples{

map_maker_random(5,5)
}
