% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_titres.R
\name{miss_titres_random}
\alias{miss_titres_random}
\title{Remove titres to create random missing data}
\usage{
miss_titres_random(titre, proportion, keep_homologous = T)
}
\arguments{
\item{titre}{A matrix of titres}

\item{proportion}{The proportion of titres to be removed}

\item{keep_homologous}{Whether to retain homologous titres}
}
\value{

}
\description{
Generates a titre table with missing data
}
\examples{
m <- map_maker_random(5, 5, 10)
ti <- dist_to_hi_titre(m$dist)
miss_titres_random(ti$lessthan_titre, 0.2)
}
